add token verify modify sauce

+++ backend/controllers/sauce.js
@@ -40,7 +40,6 @@ exports.like = (req, res, next) => {
                 sauce.dislikes++;
                 sauce.usersDisliked.push(req.body.userId);
             }
-            console.log(sauce);
             sauce
                 .save()
                 .then(() => res.status(200).json({ message: "OK" }))
@@ -61,6 +60,14 @@ exports.getOneSauce = (req, res, next) => {
 exports.modifySauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
         .then((sauce) => {
+            if (!sauce) {
+                return res.status(404).json({ error: "Sauce introuvable" });
+            }
+            if (sauce.userId !== req.auth.userId) {
+                return res
+                    .status(401)
+                    .json({ error: "Vous n'avez pas les droits" });
+            }
             const filename = sauce.imageUrl.split("/images/")[1];
             if (req.file) {
                 fs.unlink(`images/${filename}`, () => {});
